backend-env/requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.5
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
configobj==5.0.6
crypto==1.4.1
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
gast==0.3.3
h5py==2.10.0
httplib2==0.9.2
idna==2.8
importlib-metadata==1.6.1
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
language-selector==0.1
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mysqlclient==1.4.4
Naked==0.1.31
nbconvert==5.6.1
nbformat==5.0.6
nemo-emblems==4.0.2
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.4
onboard==1.4.1
opencv-python==4.1.2.30
packaging==20.3
PAM==0.4.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.4.2
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.6.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.5
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-xapp==1.4.0
python-xlib==0.20
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
scikit-learn==0.21.3
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
setproctitle==1.1.10
shellescape==3.4.1
six==1.14.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.7
virtualenv==16.7.7
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.11.2
xkit==0.0.0
zipp==3.1.0
